+++ packages/ui-react/src/hooks/useGameLogic.ts
@@ -1,4 +1,5 @@
 import {
+  BP,
   CASINO_GAME_TYPE,
   casinoChainById,
   chainById,
@@ -57,7 +58,7 @@ interface UseGameLogicResult<T extends GameChoice = GameChoice> {
   gasPrice: string
   targetPayoutAmount: bigint
   formattedNetMultiplier: number
-  grossMultiplier: number // BP
+  grossMultiplier: BP
   isInGameResultState: boolean
   isGamePaused: boolean
   nativeCurrencySymbol: string